move insertPrompt command
